+++ crates/gen/src/lib.rs
@@ -1,57 +1,62 @@
-mod gen;
-mod gen_helpers;
-mod interface_info;
-mod iterator;
-mod object;
-mod propertykey;
-mod to_ident;
+mod r#async;
 mod bool32;
 mod bstr;
 mod callback;
 mod class;
 mod com_interface;
+mod constant;
 mod delegate;
 mod r#enum;
+mod function;
+mod gen;
 mod handle;
+mod helpers;
 mod interface;
+mod interface_info;
+mod iterator;
 mod matrix3x2;
 mod matrix4x4;
 mod ntstatus;
+mod object;
+mod propertykey;
 mod pstr;
 mod pwstr;
 mod r#struct;
 mod timespan;
+mod to_ident;
 mod vector2;
 mod vector3;
 mod vector4;
-mod r#async;
 
 pub use gen::*;
-pub use gen_helpers::*;
+pub use helpers::*;
 pub use interface_info::*;
+pub use to_ident::*;
+
+use bool32::*;
+use bstr::*;
+use callback::*;
+use class::*;
+use com_interface::*;
+use constant::*;
+use delegate::*;
+use function::*;
+use handle::*;
+use interface::*;
 use iterator::*;
+use matrix3x2::*;
+use matrix4x4::*;
+use ntstatus::*;
 use object::*;
 use propertykey::*;
+use pstr::*;
+use pwstr::*;
 use quote::*;
+use r#async::*;
+use r#enum::*;
+use r#struct::*;
 use reader::*;
-pub use to_ident::*;
-pub use bool32::*;
-pub use bstr::*;
-pub use callback::*;
-pub use class::*;
-pub use com_interface::*;
-pub use delegate::*;
-pub use handle::*;
-pub use interface::*;
-pub use matrix3x2::*;
-pub use matrix4x4::*;
-pub use ntstatus::*;
-pub use pstr::*;
-pub use pwstr::*;
-pub use r#enum::*;
-pub use r#struct::*;
-pub use timespan::*;
-pub use vector2::*;
-pub use vector3::*;
-pub use vector4::*;
-pub use r#async::*;
+use timespan::*;
+use vector2::*;
+use vector3::*;
+use vector4::*;
